remove footer stories
